chores: refactor code structure to src/hicstuff and pyproject.toml and other bells and whistles
