--- .appveyor.yml
@@ -1,38 +1,0 @@
-environment:
-  # Need this to set up compilation on Windows.
-  CMD_IN_ENV: "cmd /E:ON /V:ON /C .\\appveyor\\run_with_env.cmd"
-  
-  PROJECT_NAME: hicstuff
-
-  - PIP3: "%PYTHON_VERSION%\\Scripts\\pip"
-
-  matrix:
-
-    - PYTHON: "C:\\Python34-x64"
-      PYTHON_VERSION: "3.4.x"
-      PYTHON_ARCH: "64"
-
-    - PYTHON: "C:\\Python35-x64"
-      PYTHON_VERSION: "3.5.x"
-      PYTHON_ARCH: "64"
-
-    - PYTHON: "C:\\Python36-x64"
-      PYTHON_VERSION: "3.6.x"
-      PYTHON_ARCH: "64"
-
-# We always use a 64-bit machine, but can build x86 distributions
-# with the TARGET_ARCH variable.
-platform:
-    - x64
-
-install:
-    # Clone simply to get the script for setting up Windows build environment.
-    - cmd: git clone https://github.com/pelson/Obvious-CI.git
-    - cmd: "%PIP3% install -r requirements.txt"
-
-# Skip .NET project specific build phase.
-build: off
-
-test_script:
-    # Install the package
-- pip install C:\projects\%PROJECT_NAME% pytest
