--- .github/workflows/build.yml
@@ -1,47 +1,0 @@
-name: CI
-
-on: [push]
-
-jobs:
-  ci:
-    runs-on: ubuntu-latest
-    strategy:
-      max-parallel: 5
-
-    steps:
-
-    - name: Checkout repo
-      uses: actions/checkout@v2
-    
-    - name: Create micromamba env. for package
-      uses: mamba-org/setup-micromamba@v1
-      with:
-        generate-run-shell: true
-        environment-file: environment.yml
-
-    - name: Install package
-      run: |
-        pip install .
-      shell: micromamba-shell {0}
-
-    - name: Check installed package
-      run: |
-        hicstuff --version
-      shell: micromamba-shell {0}
-
-    - name: Lint
-      run: |
-        pip install pytest-pylint pytest pytest-cov pylint codecov mappy
-        pytest --pylint --pylint-error-types=EF --pylint-rcfile=.pylintrc --doctest-modules --doctest-modules hicstuff
-      shell: micromamba-shell {0}
-
-    - name: Test
-      run: |
-        pytest --cov=hicstuff
-      shell: micromamba-shell {0}
-
-    - name: Code coverage
-      run: |
-        codecov
-      shell: micromamba-shell {0}
-
