--- .github/workflows/pypi-publish.yml
@@ -1,33 +1,0 @@
-# This workflow will upload a Python Package using Twine when a release is created
-
-name: Upload Python Package on PyPI
-
-on:
-  push:
-    tags:
-      - '*'
-
-jobs:
-  deploy:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python 3.8
-      uses: actions/setup-python@v2
-      with:
-        python-version: '3.8'
-
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install setuptools wheel twine
-        
-    - name: Build and publish
-      env:
-        TWINE_USERNAME: __token__
-        TWINE_PASSWORD: ${{secrets.PYPI_TOKEN}}
-      run: |
-        python setup.py sdist bdist_wheel
-        twine upload dist/*
