--- environment.yml
@@ -1,25 +1,0 @@
-name: hicstuff
-channels:
-  - conda-forge
-  - bioconda
-  - nodefaults
-dependencies:
-  - python >= 3.7
-  - pip
-  - bowtie2 
-  - bwa 
-  - minimap2
-  - samtools
-  - numpy
-  - scipy
-  - pandas >= 1.5.0
-  - matplotlib >= 3.4.0
-  - docopt
-  - biopython
-  - requests
-  - scikit-learn
-  - pysam
-  - htslib
-  - pyfastx
-  - cooler
-  - pairtools
