--- requirements.txt
@@ -1,12 +1,0 @@
-numpy
-scipy
-pandas >= 1.5.0
-matplotlib >= 3.4.0
-docopt
-biopython
-requests
-scikit-learn
-pysam
-pyfastx
-cooler
-pairtools
